+++ tests/models/test_several_databases.py
@@ -51,6 +51,7 @@ def test_model_with_database_name_specified(sheraf_database, db2):
     with sheraf.connection(commit=True):
         m = Db2Model.create()
         n = Db2Model.create()
+        m_id = m.id
 
     with sheraf.connection() as connection:
         assert m.id not in connection.root().get(Db2Model.table, {})
@@ -63,7 +64,7 @@ def test_model_with_database_name_specified(sheraf_database, db2):
         assert n in Db2Model.all()
 
         m.delete()
-        assert m.id not in connection_2.root()[Db2Model.table]["id"]
+        assert m_id not in connection_2.root()[Db2Model.table]["id"]
         assert Db2Model.count() == 1
 
         n.delete()
@@ -83,6 +84,7 @@ def test_database_retrocompatibility(sheraf_database, db2):
     with sheraf.connection(commit=True):
         m0 = Model.create()
         m1 = Model.create()
+        m1_id = m1.id
 
     class Model(tests.UUIDAutoModel):
         database_name = "db2"
@@ -101,7 +103,7 @@ def test_database_retrocompatibility(sheraf_database, db2):
         assert m2 in Model.all()
 
         m1.delete()
-        assert m1.id not in root1[Model.table]["id"]
+        assert m1_id not in root1[Model.table]["id"]
         assert Model.count() == 2
 
         m0.delete()